Added OSC Trigger and OSC output of API response
Python now boots as an OSC server listening for an OSC message of "/trigger" and will then transcribe an audio file. When transcription is done and the response is in from the API, it sends the response as a string through OSC using filter "/response"

I have included my TouchDesigner file that I've been testing with to show how the general flow is working. I've also included the .wav file that my TD patch is creating